split out merge strategies
